requirements.txt
# Core Framework
litestar[standard]==2.12.1
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Validation & Serialization
pydantic==2.10.3
pydantic-settings==2.7.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Logging
loguru==0.7.3

# Telegram Bot
aiogram==3.16.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1

# Utils
python-dotenv==1.0.1

# Note: Requires Python 3.13 (pydantic-core не поддерживает 3.14 пока)
